+++ packages/lib/utils/datetime.ts
@@ -43,6 +43,26 @@ export function formatDateFields<T extends z.ZodRawShape>(
   return formattedObject as z.infer<typeof zodSchema>;
 }
 
+const monthNames = [
+  "January",
+  "February",
+  "March",
+  "April",
+  "May",
+  "June",
+  "July",
+  "August",
+  "September",
+  "October",
+  "November",
+  "December",
+];
+
+// Helper function to get the month name
+export const getMonthName = (monthIndex: number) => {
+  return monthNames[monthIndex];
+};
+
 export const formatDateWithOrdinal = (date: Date): string => {
   const getOrdinalSuffix = (day: number) => {
     const suffixes = ["th", "st", "nd", "rd"];
@@ -51,20 +71,6 @@ export const formatDateWithOrdinal = (date: Date): string => {
   };
 
   const dayOfWeekNames = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
-  const monthNames = [
-    "January",
-    "February",
-    "March",
-    "April",
-    "May",
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December",
-  ];
 
   const dayOfWeek = dayOfWeekNames[date.getDay()];
   const day = date.getDate();
@@ -72,6 +78,22 @@ export const formatDateWithOrdinal = (date: Date): string => {
   const year = date.getFullYear();
 
   return `${dayOfWeek}, ${monthNames[monthIndex]} ${day}${getOrdinalSuffix(day)}, ${year}`;
+};
+
+// Helper function to format the date with an ordinal suffix
+export const getOrdinalDate = (date: number) => {
+  const j = date % 10,
+    k = date % 100;
+  if (j === 1 && k !== 11) {
+    return date + "st";
+  }
+  if (j === 2 && k !== 12) {
+    return date + "nd";
+  }
+  if (j === 3 && k !== 13) {
+    return date + "rd";
+  }
+  return date + "th";
 };
 
 export function isValidDateString(value: string) {
